# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SMARTBOSScontrol
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  brand: string;
  costPrice: number;
  salePrice: number;
  quantity: number;
  barcode?: string | null;
  createdAt: string;
}

export interface CreateProduct {
  name: string;
  brand: string;
  costPrice: number;
  salePrice: number;
  quantity: number;
  barcode?: string | null;
}

export interface DashboardStats {
  totalProducts: number;
  totalItems: number;
  totalCostValue: number;
  totalSaleValue: number;
  lowStockCount: number;
  todaySales: number;
  todayTransactions: number;
  totalDebt: number;
}

export interface SaleItem {
  id: number;
  productId?: number | null;
  productName: string;
  brand: string;
  unitPrice: number;
  quantity: number;
  totalPrice: number;
}

export interface SaleWithItems {
  id: number;
  totalAmount: number;
  itemCount: number;
  note?: string | null;
  paymentType: string;
  customerId?: number | null;
  customerName?: string | null;
  paidAmount?: number | null;
  debtAmount?: number | null;
  createdAt: string;
  items: SaleItem[];
}

export interface CreateSaleItem {
  productId: number;
  quantity: number;
}

export type CreateSalePaymentType =
  (typeof CreateSalePaymentType)[keyof typeof CreateSalePaymentType];

export const CreateSalePaymentType = {
  cash: "cash",
  card: "card",
  debt: "debt",
} as const;

export interface CreateSale {
  /** @minItems 1 */
  items: CreateSaleItem[];
  note?: string | null;
  paymentType?: CreateSalePaymentType;
  customerId?: number | null;
  paidAmount?: number | null;
}

export interface Customer {
  id: number;
  name: string;
  phone: string;
  debtLimit: number;
  totalDebt: number;
  note?: string | null;
  createdAt: string;
}

export interface CreateCustomer {
  name: string;
  phone: string;
  debtLimit?: number;
  note?: string | null;
}

export interface DebtPayment {
  id: number;
  customerId: number;
  amount: number;
  note?: string | null;
  createdAt: string;
}

export interface CreateDebtPayment {
  amount: number;
  note?: string | null;
}

export interface StatementSale {
  id: number;
  totalAmount: number;
  itemCount: number;
  paymentType: string;
  paidAmount?: number | null;
  debtAmount?: number | null;
  createdAt: string;
}

export interface StatementPayment {
  id: number;
  amount: number;
  note?: string | null;
  createdAt: string;
}

export interface CustomerStatement {
  customer: Customer;
  sales: StatementSale[];
  payments: StatementPayment[];
}

export interface NotFound {
  error: string;
}