# lib/api-zod/src/generated/types/createCustomer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SMARTBOSScontrol
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCustomer {
  name: string;
  phone: string;
  debtLimit?: number;
  note?: string | null;
}