# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SMARTBOSScontrol
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id: number;
  name: string;
  phone: string;
  debtLimit: number;
  totalDebt: number;
  note?: string | null;
  createdAt: string;
}