# lib/api-zod/src/generated/types/customerStatement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SMARTBOSScontrol
 * OpenAPI spec version: 0.1.0
 */
import type { Customer } from "./customer";
import type { StatementPayment } from "./statementPayment";
import type { StatementSale } from "./statementSale";

export interface CustomerStatement {
  customer: Customer;
  sales: StatementSale[];
  payments: StatementPayment[];
}