# lib/api-zod/src/generated/types/debtPayment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SMARTBOSScontrol
 * OpenAPI spec version: 0.1.0
 */

export interface DebtPayment {
  id: number;
  customerId: number;
  amount: number;
  note?: string | null;
  createdAt: string;
}