# lib/api-zod/src/generated/types/statementPayment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SMARTBOSScontrol
 * OpenAPI spec version: 0.1.0
 */

export interface StatementPayment {
  id: number;
  amount: number;
  note?: string | null;
  createdAt: string;
}