# lib/api-zod/src/generated/types/statementSale.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SMARTBOSScontrol
 * OpenAPI spec version: 0.1.0
 */

export interface StatementSale {
  id: number;
  totalAmount: number;
  itemCount: number;
  paymentType: string;
  paidAmount?: number | null;
  debtAmount?: number | null;
  createdAt: string;
}